threatnote/requirements.txt
OTXv2
flask
flask_login
redis
flask_wtf
requests
rq
whois
virustotal-api
flask_sqlalchemy
python-dotenv
shodan
redis
flask_migrate
flask_dotenv
gunicorn
markdown2
pymisp
flask_weasyprint
